+++ event_sourcing/test/event_store/event_hash_versions_test.dart
@@ -5,7 +5,6 @@ import 'dart:convert';
 
 import 'package:crypto/crypto.dart';
 import 'package:event_sourcing/event_sourcing.dart';
-import 'package:event_sourcing/src/storage/event_hash.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sembast/sembast_memory.dart';