+++ event_sourcing/test/storage/storage_backend_conformance.dart
@@ -30,7 +30,6 @@ import 'dart:async';
 
 import 'package:event_sourcing/event_sourcing.dart';
 import 'package:event_sourcing/src/security/security_context_store.dart';
-import 'package:event_sourcing/src/storage/event_hash.dart';
 import 'package:event_sourcing/src/testing/delivery_test_hooks.dart';
 import 'package:flutter_test/flutter_test.dart';